Add an MPRIS plugin
The Media Player Remote Interfacing Specification (MPRIS) is a standard
D-Bus interface which aims to provide a common programmatic API for
controlling media players. It's widely supported and implemented by the
majority of media players, as well as browsers, desktops, etc.

While we used to have MPRIS as part of core, it was an implementation of
an earlier specification, and the code got stale and unused before it
got eventually removed.

This is a new implementation written from scratch and targeting v2.2 of
the specification and specifically the org.mpris.MediaPlayer2 and
org.mpris.MediaPlayer2.Player interfaces.

The implementation is leveraging the dbus-next Python library, as the
MPRIS specification is relying heavily on Properties, which the dbus
library has no support for. This means that we now depend on two D-Bus
libraries, which is unfortunate, but out of scope for this commit. The
existing D-Bus code (DBusFacade, GNOME Media Keys, MATE Media Keys) can
be easily ported over to dbus-next if required, although it should be
explored whether MPRIS support supersedes them.